requirements.txt
statsmodels
pandas
pymapd
darkskylib
requests
geopy
tensorflow-gpu
numpy
beautifulsoup4
darkskylib
scikit_learn
configparser
dask